+++ plugin-api/src/main/java/org/sonar/api/issue/impact/Severity.java
@@ -17,48 +17,10 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.api.batch.sensor.issue;
+package org.sonar.api.issue.impact;
 
-import java.util.Map;
-import javax.annotation.CheckForNull;
-import org.sonar.api.batch.rule.Severity;
-import org.sonar.api.batch.sensor.Sensor;
-import org.sonar.api.issue.impact.SoftwareQuality;
-import org.sonar.api.rules.RuleType;
-
-/**
- * Represents an issue imported from an external rule engine by a {@link Sensor}.
- * @since 7.2
- */
-public interface ExternalIssue extends IIssue {
-
-  /**
-   * @since 7.4
-   */
-  String engineId();
-
-  /**
-   * @since 7.4
-   */
-  String ruleId();
-
-  Severity severity();
-
-  /**
-   * Effort to fix the issue, in minutes.
-   */
-  @CheckForNull
-  Long remediationEffort();
-
-  /**
-   * Type of the issue.
-   */
-  RuleType type();
-
-  /**
-   * Impacts of the issue.
-   * @since 10.1
-   */
-  Map<SoftwareQuality, org.sonar.api.issue.impact.Severity> impacts();
-
+public enum Severity {
+  LOW,
+  MEDIUM,
+  HIGH
 }