+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 10.1
 
 * Remove @Beta Code Characteristics `org.sonar.api.code.CodeCharacteristic`
+* Introduce `org.sonar.api.issue.impact.SoftwareQuality` and `org.sonar.api.issue.impact.Severity` to define impacts of rules and issues
 
 ## 10.0